refactor(android): extract ModeContentHost + Resource sealed interface + shared StringChunking
- ModeContentHost (Step 1.11) — one-line dispatch swaps per mode, reduces merge surface
- Resource<T> (R8) — common ui/common/Resource.kt sealed interface used by Wave 2+ ViewModels
- StringChunking (R4) — moves chunkDocument from DocumentScreen to ui/common so Stats/Doctor renderers can share
